third_party/WebKit/LayoutTests/external/wpt/html/semantics/interactive-elements/the-dialog-element/dialog-autofocus-multiple-times.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/common.js"></script>
<script>
var onloadedAndFocusedTest = async_test("on-loaded-And-Autofocused");

function onLoadedAndFocused() {
    onloadedAndFocusedTest.step(function() {
        assert_equals(document.activeElement, document.getElementById("outer-button"));

        var focusCount = 0;
        input2.onfocus = function() { focusCount += 1 };

        var expectedFocusCount = 3;
        for (i = 0; i < expectedFocusCount; i++) {
            dlg.show();
            assert_equals(document.activeElement, document.getElementById("input2"));
            input1.focus();
            assert_equals(document.activeElement, document.getElementById("input1"));
            dlg.close();
        }

        assert_equals(focusCount.toString(), expectedFocusCount.toString());
    });
    onloadedAndFocusedTest.done();
}

test(function() {
  waitUntilLoadedAndAutofocused(onLoadedAndFocused);
}, "Tests autofocus is run every time a dialog is opened.");
</script>
</head>
<body>
<button id="outer-button" autofocus></button>
<dialog id="dlg">
    <!-- Unfocusable elements with [autofocus] should be ignored. -->
    <input autofocus disabled>
    <textarea autofocus hidden></textarea>
    <input id="input1"></input>
    <input id="input2" autofocus></input>
</dialog>
</body>
</html>
